Minor bug fixes and general clean-up.

+++ CASpeex/src/CAOggSpeexDecoder.cpp
@@ -102,15 +102,15 @@ void CAOggSpeexDecoder::SetCurrentInputFormat(const AudioStreamBasicDescription&
 UInt32 CAOggSpeexDecoder::ProduceOutputPackets(void* outOutputData, UInt32& ioOutputDataByteSize, UInt32& ioNumberPackets,
                                                AudioStreamPacketDescription* outPacketDescription)
 {
-    dbg_printf(" >> [%08lx] CAOggSpeexDecoder :: ProduceOutputPackets(%ld [%ld])\n", (size_t) this, ioNumberPackets, ioOutputDataByteSize);
+    dbg_printf(" >> [%08lx] CAOggSpeexDecoder :: ProduceOutputPackets(%u [%u])\n", (size_t) this, (unsigned int)ioNumberPackets, (unsigned int)ioOutputDataByteSize);
     UInt32 ret = kAudioCodecProduceOutputPacketSuccess;
 
     if (mFramesBufferedList.empty()) {
         ioOutputDataByteSize = 0;
         ioNumberPackets = 0;
         ret = kAudioCodecProduceOutputPacketNeedsMoreInputData;
-        dbg_printf("<!E [%08lx] CAOggSpeexDecoder :: ProduceOutputPackets(%ld [%ld]) = %ld [%ld]\n", (size_t) this,
-                   ioNumberPackets, ioOutputDataByteSize, ret, FramesReady());
+        dbg_printf("<!E [%08lx] CAOggSpeexDecoder :: ProduceOutputPackets(%u [%u]) = %u [%u]\n", (size_t) this,
+                   (unsigned int)ioNumberPackets, (unsigned int)ioOutputDataByteSize, (unsigned int)ret, (unsigned int)(FramesReady()));
         return ret;
     }
 
@@ -249,7 +249,7 @@ void CAOggSpeexDecoder::InPacket(const void* inInputData, const AudioStreamPacke
 
     mFramesBufferedList.push_back(packet_count);
 
-    dbg_printf("<.. [%08lx] CAOggSpeexDecoder :: InPacket(): packet_count: %ld\n", (size_t) this, packet_count);
+    dbg_printf("<.. [%08lx] CAOggSpeexDecoder :: InPacket(): packet_count: %u\n", (size_t) this, (unsigned int)packet_count);
 }
